+++ docs/src/plugin-pinia/changelog.md
@@ -10,13 +10,17 @@ outline: 2
 
 ## Next
 
+### Features
+
+- Add experimental support for store migrations.
+
 ### Bug Fixes
 
 - Fix panic when calling [`ManagerExt::pinia`](https://docs.rs/tauri-plugin-pinia/3.0.0/tauri_plugin_pinia/trait.ManagerExt.html#method.pinia) from [`Window`](https://docs.rs/tauri/2.3.1/tauri/window/struct.Window.html) or [`WebviewWindow`](https://docs.rs/tauri/2.3.1/tauri/webview/struct.WebviewWindow.html).
 
 ### Dependencies
 
-- Update `tauri-store` to `0.9.1`.
+- Update `tauri-store` to `0.10.0`.
 
 ## 3.0.0